Use i64 to store format in ReadError::InvalidFormat
This lets us store the majority of format types we might encounter.

+++ font-tables/src/read.rs
@@ -49,7 +49,8 @@ pub trait ComputeSize: ReadArgs {
 #[derive(Debug, Clone)]
 pub enum ReadError {
     OutOfBounds,
-    InvalidFormat(u16),
+    // i64 is flexible enough to store any value we might encounter
+    InvalidFormat(i64),
     InvalidSfnt(u32),
     InvalidArrayLen,
     ValidationError,
